Fix orange to be official Bitcoin logo orange
The orange previously was a bit darker because I thought that this would
be necessary for the QR code to scan on popular Android scanners.

After more testing, I noticed that whether the code scans depends
mostly on the color of the surroundings and ambient light.

Thus, this commit changes the orange to be the "official" one.